Affine DOF and Force sensor testing
Added Force sensor definitions to robot definition.
Added a separate "fullDOF" robot for planning purposes. This robot
definition is for planning only, as the affine DOF's mess with the
odevelocity controller. Ideally, the planner would work directly with
affine DOF, but so far this is not the case.
TODO: complete FT sensor definitions in main robot file
TODO: test sensor output in ankles and hands